# docs/context/engineering/api-contracts/v1/openapi.yaml
openapi: 3.1.0
info:
  title: InComedy Server API
  version: "1.0.0"
  description: |
    Current backend API contract for auth, identity foundation, organizer workspace foundation,
    and service health endpoints.
servers:
  - url: https://incomedy.ru
    description: Production
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a first-party credential account and issue a session
      operationId: postAuthRegister
      parameters:
        - in: header
          name: X-Request-ID
          required: false
          description: Optional client-provided UUID request id for tracing.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialAuthRequest"
      responses:
        "200":
          description: Account registered and session issued successfully
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionResponse"
        "400":
          description: Invalid login/password payload
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Login is already registered
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Request body exceeds route size limit
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Sign in with login and password and issue a session
      operationId: postAuthLogin
      parameters:
        - in: header
          name: X-Request-ID
          required: false
          description: Optional client-provided UUID request id for tracing.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialAuthRequest"
      responses:
        "200":
          description: Session issued successfully
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionResponse"
        "400":
          description: Invalid login/password payload
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid login or password
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Request body exceeds route size limit
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    CredentialAuthRequest:
      type: object
      additionalProperties: false
      required:
        - login
        - password
      properties:
        login:
          type: string
          minLength: 3
          maxLength: 32
        password:
          type: string
          minLength: 8
          maxLength: 128
...
    AuthSessionResponse:
      type: object
      additionalProperties: false
      required:
        - provider
        - access_token
        - refresh_token
        - expires_in
        - user
      properties:
        provider:
          type: string
          enum: [password, phone, vk, telegram, google, apple]
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          format: int64
          example: 3600
        user:
          $ref: "#/components/schemas/SessionUser"
...
    SessionUser:
      type: object
      additionalProperties: false
      required:
        - id
        - display_name
        - roles
        - linked_providers
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
          example: John Doe
        username:
          type: string
          nullable: true
        photo_url:
          type: string
          nullable: true
        roles:
          type: array
          items:
            type: string
            enum: [audience, comedian, organizer]
        active_role:
          type: string
          nullable: true
          enum: [audience, comedian, organizer]
        linked_providers:
          type: array
          items:
            type: string
            enum: [password, phone, vk, telegram, google, apple]
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: auth_failed
        message:
          type: string
          example: Authentication failed